requirements.txt
# API and LLM Integration
anthropic>=0.7.0     # Claude API client for AI tutoring capabilities
python-dotenv>=1.0.0 # Environment variable management for API keys

# Knowledge Representation
owlready2>=0.45      # Ontology management and reasoning

# Web Framework  
flask>=2.3.0         # Web server
flask-cors>=4.0.0    # Cross-origin resource sharing support
serverless-wsgi>=3.0.1  # WSGI support for serverless deployment

# Security
python-jose[cryptography]>=3.3.0  # JWT token handling
bcrypt>=4.0.1        # Password hashing
python-multipart>=0.0.6  # Form data handling

# Natural Language Processing
nltk>=3.8.1          # Natural language toolkit for text processing 